memory-algebra: make the cost model produce numbers a calibration study can use
Phase 0 of the calibration plan (memory-algebra-paper/EXPERIMENTS.md) asks for property
tests and a prediction extractor. Auditing first found the model could not support
either: on the reference fleet, the same 64 KiB tile predicted `None` for the
host<->device seam and `1` for every on-die `B/s` hop, on every SKU. Five defects, all
of which would have been charged to hardware once measurement began.

1. No time resolution. `derived_transfer_cost` did `bytes.div_ceil(bw.bytes)` with
   `bw.bytes` in bytes-per-SECOND, so any transfer under one second cost 1 -- which is
   every transfer these machines perform. Swept 4 KiB..64 MiB: a 16,384x range in bytes
   gave the identical prediction, and a 2.4x spread in declared L2 bandwidth across the
   fleet gave no difference at all. Time costs are now picoseconds; the fleet's HBM->L2
   spans 311 ps to 317,892 ps over that sweep, and H100/A100 differ by 1.714x = 12/7
   exactly, which is the ratio of the numbers the files declare.

2. Units discarded at the harvest boundary. `DerivedCost` carried `RatePer`;
   `StagingRoute.derived_cost` was a bare `Option<u32>`. The fleet's `HBM->L2` (seconds)
   and `L2->SMEM` (cycles) therefore appeared in one JSON column as `1` and `128` --
   different dimensions, nothing recording which. Now `derived_unit: "ps"|"cyc"`, and
   the cost is `u64` through AST/HIR and the `vx.transfer` attribute (i32 overflows at
   4.3 ms of picoseconds).

3. A 9.05% systematic error in every bandwidth. `TB` parsed as 2^40 while every `spec:`
   line in fleet/ cites a vendor decimal figure -- NVIDIA's "3.35 TB/s" is 3.35e12 B/s
   from a 5120-bit bus at 3.2 Gbps. The compiler read each link ~10% faster than its own
   citation and understated every predicted time to match. New `src/units.rs` is the one
   place a declared figure becomes bytes: SI decimal, IEC binary, both spellings
   accepted, exact u128 integer arithmetic, and a quantity that is not a whole number of
   bytes is rejected rather than rounded (`1.1 KiB` is an error, not 1126). `Mb` does not
   parse as `MB`.

   Fleet capacities/granules were respelled to IEC, which preserves their byte values
   exactly -- `80 GiB` is the same 85899345920 as `80 GB` was -- so no admission verdict
   could move. Verified: `capacity_bytes` is byte-identical, and admit.vx still rejects
   on H100 (E6009/E6010, 120 GiB working set vs 80 GiB) and admits on B200/MI300X.

4. Two cost sources per edge, silently. Every fleet file declared
   `transfer HBM -> L2 : 40` *and* gave L2 a bandwidth, so the edge had a declared cost
   and a derivable one; the compiler routed by the first and reported the second, and
   nothing detected the disagreement. `EdgeCost` is now `Derived | Rate(Bandwidth) |
   Fixed(u32)`, the `: cost` is optional, and E6013 rejects an edge that has both.
   Resolving it by precedence was the alternative and is worse: it silently discards a
   figure someone wrote down on purpose. (vx-review#10, the S-series blocker.)

5. Machine-file topologies were never coherence-checked. `check_topology_coherence` read
   `ast.topologies` -- the entry program alone -- while `check_memory_coherence` reads the
   env. So every fleet SKU had its memory declarations validated and its topology
   declarations skipped: E6005, W1026, W1027 and now E6013 had never run against the
   files they exist for. Reading the env instead is what surfaced (4) on all seven SKUs.

With those fixed, `EdgeCost::Rate` gives the host seam the prediction it never had: a
link bandwidth belongs to PCIe or C2C, not to either memory it connects, so containment
cannot derive it and the roofline walk returned nothing. Declared per SKU with `spec:`
provenance and an explicit UNVERIFIED marker naming M1/M4 as the experiment that settles
it -- a Grace/C2C-attached part reaches the host ~7x faster than PCIe Gen5, and guessing
wrong is exactly the kind of error this pass exists to stop. The seam now predicts
260,064 ps on Gen5 parts and 520,127 ps on the Gen4 A100s: a 2.0x ratio, which is 63/31.5.

Note for #14: every prediction harvested before this commit is stale.

Ref: vx-review#23, vx-review#10

+++ src/hir/arena.rs
@@ -84,7 +84,9 @@ pub struct HirSpawnOnExpr {
 pub struct HirTransferExpr {
     pub expr: ExprId,
     pub space: MemorySpace,
-    pub cost: Option<u32>,
+    /// Roofline cost: cycles for a `B/cyc` hop, picoseconds for a `B/s` one. See
+    /// [`crate::hir::memory::DerivedCost`].
+    pub cost: Option<u64>,
     pub span: Span,
 }
